services/ingest-meets/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
streamlit>=1.32.0
plotly>=5.20.0
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.43.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.3.0
rapidfuzz>=3.8.0
apscheduler>=3.10.0
python-multipart>=0.0.9
pydantic>=2.6.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
